refactor: Split restore cutoff helpers from copy flow
issue: #50663

Restore cutoff rewrite logic had grown inside the generic segment copy
helper, which made the normal copy path harder to review and mixed two
separate responsibilities in one file. Move the cutoff-specific insert,
delta, manifest, and record helpers into a dedicated file, and move the
corresponding tests with them.

This keeps the copy path behavior unchanged while reducing the size of
the generic copy helper and making future cutoff review more local.

Validation:
- git diff --check
- go test -tags dynamic,test -gcflags="all=-N -l" \
  github.com/milvus-io/milvus/internal/datanode/importv2 \
  -run 'Test(BuildCutoffTextColumnConfigs_UsesRewriteMode|RewriteV2InsertBinlogsForCutoff_RewritesCrossingBatch|RewriteManifestSegmentForCutoff|WriteAndScanDeltaLog|CutoffHelpers|CopySegmentAndIndexFiles_V2WithCutoff|CopySegmentAndIndexFiles_L0WithCutoff)' \
  -v -count=1
- go test -tags dynamic,test -gcflags="all=-N -l" \
  github.com/milvus-io/milvus/internal/datacoord \
  -run 'Test(IsSnapshotSegmentCandidate_IncludesSegmentAtSnapshotTs|GenSnapshot_CommitTimestamp)' \
  -v -count=1
